refactor: runtime.py 拆出工具路由策略层 (C10)
- 新增 app/agent/tool_routing.py：浏览器页面模式/可见浏览器模式/
  Windows 控制/屏幕观察的工具过滤与提示词规则（26 个纯函数，467 行）
- runtime.py re-export 全部拆出符号，旧 import 路径零变化；
  瘦身 2246→1648 行，核心工具循环不动
- 与 runtime 模型消息构建层的两处交叉引用改为函数内延迟 import 避环

测试：零测试改动；runtime 单元+集成 162 passed；全量 813 passed / exit 0
